pyproject.toml
[tool.poetry]
name = "remind-knowledge-populating-worker"
version = "0.1.0"
description = "A RabbitMQ worker for populating knowledge in the ReMind system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "worker"}]

[tool.poetry.dependencies]
python = "^3.9"
pika = "^1.3.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pydantic = "^2.5.0"
openai = "^1.12.0"
google-generativeai = "^0.3.0"
langgraph = "^0.2.0"
langchain = "^0.2.0"
langchain-openai = "^0.1.0"
langchain-community = "^0.2.0"
chromadb = "^0.4.0"
psycopg2-binary = "^2.9.10"
google-adk = "^1.8.0"
pre-commit = "^4.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
flake8 = "^6.0.0"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
